update balance type from float64 to int64

+++ internal/entity/account.go
@@ -7,7 +7,7 @@ type Account struct {
 	Username     string    `db:"username"`
 	PasswordHash string    `db:"password_hash"`
 	IsAdmin      bool      `db:"is_admin"`
-	Balance      float64   `db:"balance"`
+	Balance      int64     `db:"balance"`
 	CreatedAt    time.Time `db:"created_at"`
 	UpdatedAt    time.Time `db:"updated_at"`
 }
